Crate trillium_static_compiled_macros

Internal-use fork of include_dir_macros for trillium.rs. It is not intended for general use. Credit for the bulk of the code goes to the authors of the upstream crate.

Differences from upstream:

include_entry was added, which returns a DirEntry instead of a Dir, making direct inclusion of files possible

Metadata is always enabled

relative paths are resolved from a root of CARGO_MANIFEST_DIR

hygiene is maintained by using a macro_rules macro to import relevant structs

Macros

include_dir

Embed the contents of a directory. “Returns” a Dir

include_entry

Embed a directory or file. “Returns” a DirEntry